Promisify all library functions
