+++ megaphone-validator/lib/reupload-audio.js
@@ -5,21 +5,25 @@ const authorizationHeader = {
 }
 
 module.exports = (megaphoneEpisode, audioUrl) => {
-    const episodeId = megaphoneEpisode.id;
-    const networkId = process.env.MEGAPHONE_NETWORK_ID;
-    const podcastId = megaphoneEpisode.podcastId;
-    const options = {
-        body: {
-            backgroundAudioFileUrl: audioUrl
-        },
-        headers: authorizationHeader,
-        json: true,
-        method: 'PUT',
-        url: `https://cms.megaphone.fm/api/networks/${networkId}/podcasts/${podcastId}/episodes/${episodeId}`,
-    };
+    return new Promise((resolve, reject) => {
+        const episodeId = megaphoneEpisode.id;
+        const networkId = process.env.MEGAPHONE_NETWORK_ID;
+        const podcastId = megaphoneEpisode.podcastId;
+        const options = {
+            body: {
+                backgroundAudioFileUrl: audioUrl
+            },
+            headers: authorizationHeader,
+            json: true,
+            method: 'PUT',
+            url: `https://cms.megaphone.fm/api/networks/${networkId}/podcasts/${podcastId}/episodes/${episodeId}`,
+        };
 
-    request(options, (err, response, body) => {
-        console.log(body);
-        console.log(body.audioFileProcessing);
+        request(options, (err, response, body) => {
+            if (err) reject(err);
+            console.log(body);
+            console.log(body.audioFileProcessing);
+            resolve(body);
+        });
     });
 };
